Finished Fit export as DataFrame
