+++ pyproject.toml
@@ -53,6 +53,8 @@ pytest-clarity = "^1.0.1"
 pytest-cov = "^2.12.1"
 pytest-console-scripts = "^1.2.0"
 pre-commit = "^2.13.0"
+types-toml = "^0.1.3"
+types-PyYAML = "^5.4.3"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
