+++ spectrafit/spectrafit.py
@@ -2,11 +2,9 @@ import argparse
 import json
 import pprint
 
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Tuple
 
-import matplotlib.pylab as plt
 import numpy as np
 import pandas as pd
 import toml
@@ -17,21 +15,10 @@ from lmfit import Parameters
 from lmfit import conf_interval
 from lmfit import report_ci
 from lmfit import report_fit
-from matplotlib.ticker import AutoMinorLocator
-from matplotlib.widgets import Cursor
-from scipy.special import erf
-from scipy.special import wofz
 from spectrafit import __version__
-from spectrafit.export import fit_report_as_dict
-
-
-@dataclass(frozen=True)
-class Constants:
-    log2 = np.log(2.0)
-    sq2pi = np.sqrt(2.0 * np.pi)
-    sqpi = np.sqrt(np.pi)
-    sq2 = np.sqrt(2.0)
-    sig2fwhm = 2.0 * np.sqrt(2.0 * np.log(2.0))
+from spectrafit.models import calculated_models
+from spectrafit.models import solver_model
+from spectrafit.report import fit_report_as_dict
 
 
 pp = pprint.PrettyPrinter(indent=4)
@@ -212,7 +199,8 @@ def command_line_runner(args: dict = None) -> None:
             return
         elif again == "y":
             print("Lets start fitting ...")
-            fitting_routine(df=df, args=args)
+            df_result = fitting_routine(df=df, args=args)
+            lin_cor = df_result.corr()
         else:
             print('You should enter either "y" or "n".')
 
@@ -242,19 +230,18 @@ def extracted_from_command_line_runner() -> dict:
                 "min_correl": 0.1,
                 "sort_pars": False,
             }
-        if "conf_interfal" in _args["fitting"]["parameters"].keys():
-            result["conf_interfal"] = _args["fitting"]["parameters"]["conf_interfal"]
+        if "conf_interval" in _args["fitting"]["parameters"].keys():
+            result["conf_interval"] = _args["fitting"]["parameters"]["conf_interval"]
         else:
-            result["conf_interfal"] = {"maxiter": 200}
+            result["conf_interval"] = None
     if "peaks" in _args["fitting"].keys():
         result["peaks"] = _args["fitting"]["peaks"]
     return result
 
 
-def fitting_routine(df: pd.DataFrame, args: dict) -> None:
+def fitting_routine(df: pd.DataFrame, args: dict) -> Tuple[pd.DataFrame, dict]:
 
     # try:
-    df_dict = df.to_dict()
     df = energy_range(df=df, args=args)
     df = energy_shift(df=df, args=args)
     df = oversampling(df=df, args=args)
@@ -262,7 +249,7 @@ def fitting_routine(df: pd.DataFrame, args: dict) -> None:
 
     params = get_parameters(args=args)
     mini = Minimizer(
-        model,
+        solver_model,
         params,
         fcn_args=(df[args["column"][0]].values, df[args["column"][1]].values),
         **args["minimizer"],
@@ -271,28 +258,31 @@ def fitting_routine(df: pd.DataFrame, args: dict) -> None:
     # try:
     result = mini.minimize(**args["optimizer"])
     fit_insights = fit_report_as_dict(result, modelpars=result.params)
-    confidence_interval = conf_interval(
-        mini, result, trace=True, **args["conf_interfal"]
+    if args["conf_interval"]:
+        confidence_interval = conf_interval(mini, result, **args["conf_interval"])
+    df = df.rename(
+        columns={args["column"][0]: "energy", args["column"][1]: "intensity"}
     )
-    df["energy"] = df[args["column"][0]].values
-    df["intensity"] = df[args["column"][1]].values
     df["residual"] = result.residual
-    df["fit"] = df[args["column"][1]].values - result.residual
+    df["fit"] = df["intensity"].values - result.residual
+    df = calculated_models(params=result.params, x=df["energy"].values, df=df)
 
     if args["verbose"]:
         print("Input Parameter:\n")
         pp.pprint(args)
         print("\nFit Results and Insights:\n")
         pp.pprint(fit_insights)
-        print("\nConfidence Interval:\n")
-        pp.pprint(confidence_interval)
+        if args["conf_interval"]:
+            print("\nConfidence Interval:\n")
+            pp.pprint(confidence_interval)
     else:
-        print(" Fit-Report\n")
+        print("\nFit Results and Insights:\n")
         print(report_fit(result, modelpars=result.params, **args["report"]))
-        if args["conf_interfal"]:
-            print(" Fit-Report\n")
-            print(report_ci(conf_interval(mini, result, **args["conf_interfal"])))
+        if args["conf_interval"]:
+            print("\nConfidence Interval:\n")
+            report_ci(conf_interval(mini, result, **args["conf_interval"])[0])
 
+    return df
     # except IOError:
     #    print "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
     #    print "!!!!No Inputfile guess.parm for Fits!!!!"
@@ -317,172 +307,6 @@ def get_parameters(args: dict) -> dict:
             for key_3, value_3 in value_2.items():
                 params.add(f"{key_2}_{key_3}_{key_1}", **value_3)
     return params
-
-
-def model(params: dict, x: np.array, data):
-    """[summary]
-
-    Args:
-        params (dict): [description]
-        x (np.array): [description]
-        data ([type]): [description]
-
-    Returns:
-        [type]: [description]
-    """
-    val = 0.0
-    for mode in params:
-        mode = mode.lower()
-        if mode.split("_")[0] in [
-            "gaussian",
-            "lorentzian",
-            "voigt",
-            "pseudovoigt",
-            "exponential",
-            "linear",
-            "constant",
-            "erf",
-            "atan",
-            "log",
-        ]:
-            pass
-        else:
-            raise SystemExit(f"{mode} is not supported")
-    for mode in params:
-        mode = mode.lower()
-        if "gaussian" in mode:
-            if "center" in mode:
-                center = params[mode]
-            if "amplitude" in mode:
-                amplitude = params[mode]
-            if "fwhm_gaussian" in mode:
-                fwhm_gaussian = params[mode]
-                val += gaussian(
-                    x=x,
-                    amplitude=amplitude,
-                    center=center,
-                    fwhm=fwhm_gaussian,
-                )
-        if "lorentzian" in mode:
-            if "center" in mode:
-                center = params[mode]
-            if "amplitude" in mode:
-                amplitude = params[mode]
-            if "fwhm_lorentzian" in mode:
-                fwhm_lorentzian = params[mode]
-                val += lorentzian(
-                    x=x,
-                    amplitude=amplitude,
-                    center=center,
-                    fwhm=fwhm_lorentzian,
-                )
-        if "voigt" in mode:
-            if "center" in mode:
-                center = params[mode]
-            if "fwhm" in mode:
-                fwhm = params[mode]
-            if "gamma" in mode:
-                gamma = params[mode]
-                val += voigt(
-                    x=x,
-                    center=center,
-                    fwhm=fwhm,
-                    gamma=gamma,
-                )
-        if "pseudovoigt" in mode:
-            if "center" in mode:
-                center = params[mode]
-            if "amplitude" in mode:
-                amplitude = params[mode]
-            if "fwhm_gaussian" in mode:
-                fwhm_gaussian = params[mode]
-            if "fwhm_lorentzian" in mode:
-                fwhm_lorentzian = params[mode]
-                val += pseudovoigt(
-                    x=x,
-                    amplitude=amplitude,
-                    center=center,
-                    fwhm_g=fwhm_gaussian,
-                    fwhm_l=fwhm_lorentzian,
-                )
-        if "exponential" in mode:
-            if "amplitude" in mode:
-                amplitude = params[mode]
-            if "decay" in mode:
-                decay = params[mode]
-            if "intercept" in mode:
-                intercept = params[mode]
-                val += exponential(
-                    x=x,
-                    amplitude=amplitude,
-                    decay=decay,
-                    intercept=intercept,
-                )
-        if "power" in mode:
-            if "amplitude" in mode:
-                amplitude = params[mode]
-            if "exponent" in mode:
-                exponent = params[mode]
-            if "intercept" in mode:
-                intercept = params[mode]
-                val += powerlaw(
-                    x=x,
-                    amplitude=amplitude,
-                    exponent=exponent,
-                    intercept=intercept,
-                )
-        if "linear" in mode:
-            if "slope" in mode:
-                slope = params[mode]
-            if "intercept" in mode:
-                intercept = params[mode]
-                val += linear(x=x, slope=slope, intercept=intercept)
-        if "constant" in mode and "amplitude" in mode:
-            amplitude = params[mode]
-            val += constant(x=x, amplitude=amplitude)
-        if "erf" in mode:
-            if "center" in mode:
-                center = params[mode]
-            if "sigma" in mode:
-                sigma = params[mode]
-            if "amplitude" in mode:
-                amplitude = params[mode]
-                val += step(
-                    x=x,
-                    amplitude=amplitude,
-                    center=center,
-                    sigma=sigma,
-                    form="erf",
-                )
-        if "atan" in mode:
-            if "center" in mode:
-                center = params[mode]
-            if "sigma" in mode:
-                sigma = params[mode]
-            if "amplitude" in mode:
-                amplitude = params[mode]
-                val += step(
-                    x=x,
-                    amplitude=amplitude,
-                    center=center,
-                    sigma=sigma,
-                    form="atan",
-                )
-        if "log" in mode:
-            if "center" in mode:
-                center = params[mode]
-            if "sigma" in mode:
-                sigma = params[mode]
-            if "amplitude" in mode:
-                amplitude = params[mode]
-                val += step(
-                    x=x,
-                    amplitude=amplitude,
-                    center=center,
-                    sigma=sigma,
-                    form="logistic",
-                )
-    return val - data
 
 
 def energy_range(df: pd.DataFrame, args: dict) -> pd.DataFrame:
@@ -537,798 +361,6 @@ def intensity_smooth(df: pd.DataFrame, args: dict) -> pd.DataFrame:
         )
         return df
     return df
-
-
-def plot(x, y, final, result, args):
-    filename = args.outfile
-    # try:
-    fig, ax = plt.subplots()
-    out = [x, y, final, y - final]
-    # tmp = open(filename+'_Export.txt','w+')
-
-    df = panda_h()
-    print("--------------")
-    print("|Fit-Results:|")
-    print("--------------")
-    # printh(tmp)
-
-    index = 1
-    for i, mode in enumerate(result.params):
-        mode = mode.lower()
-        if "gaus" in mode:
-            if "cen" in mode:
-                cen = result.params[mode]
-            if "amp" in mode:
-                amp = result.params[mode]
-            if "fwg" in mode:
-                fwg = result.params[mode]
-                val = np.nan_to_num(gaussian(x, amp, cen, fwg))
-                plt.plot(x, val, color="C9", ls=":")
-                out.append(val)
-                # Exporting
-                model = "Gaussian"
-                res = np.array([val, x])
-
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=cen,
-                    amp=amp,
-                    fwg=fwg,
-                    fwl=None,
-                    fwh=None,
-                    gam=None,
-                    dec=None,
-                    exp=None,
-                    slp=None,
-                    con=None,
-                    sig=None,
-                )
-                index += 1
-                # printf(tmp,x, val,'Gaussian',index,cen,amp,fwg,0,0,0,0,0,0,0,0)
-
-        if "lorz" in mode:
-            if "cen" in mode:
-                cen = result.params[mode]
-            if "amp" in mode:
-                amp = result.params[mode]
-            if "fwl" in mode:
-                fwl = result.params[mode]
-                val = np.nan_to_num(lorentzian(x, amp, cen, fwl))
-                plt.plot(x, val, color="C9", ls=":")
-                out.append(val)
-                # Exporting
-                model = "Lorentzian"
-                res = np.array([val, x])
-
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=cen,
-                    amp=amp,
-                    fwg=None,
-                    fwl=fwl,
-                    fwh=None,
-                    gam=None,
-                    dec=None,
-                    exp=None,
-                    slp=None,
-                    con=None,
-                    sig=None,
-                )
-                index += 1
-                # printf(tmp,x, val,'Lorentzian',index,cen,amp,0,fwl,0,0,0,0,0,0,0)
-
-        if "voigt" in mode:
-            if "cen" in mode:
-                cen = result.params[mode]
-            if "fwh" in mode:
-                fwh = result.params[mode]
-            if "gam" in mode:
-                gam = result.params[mode]
-                val = np.nan_to_num(voigt(x, cen, fwh, gam))
-                plt.plot(x, val, color="C9", ls=":")
-                out.append(val)
-                # Exporting
-                model = "Voigt"
-                res = np.array([val, x])
-                df = panda_e(
-                    df,
-                    model,
-                    res,
-                    cen=cen,
-                    amp=None,
-                    fwg=None,
-                    fwl=None,
-                    fwh=fwh,
-                    gam=gam,
-                    dec=None,
-                    exp=None,
-                    slp=None,
-                    con=None,
-                    sig=None,
-                )
-                index += 1
-                # printf(tmp,x, val,'Voigt',index,cen,amp,0,0,fwh=fwh,gam=gamma,0,0,0,0,0)
-
-        if "pseudovoigt" in mode:
-            if "cen" in mode:
-                cen = result.params[mode]
-            if "amp" in mode:
-                amp = result.params[mode]
-            if "fwg" in mode:
-                fwg = result.params[mode]
-            if "fwl" in mode:
-                fwl = result.params[mode]
-                val = np.nan_to_num(pseudovoigt(x, amp, cen, fwg, fwl))
-                plt.plot(x, val, color="C9", ls=":")
-                out.append(val)
-                # Exporting
-                model = "Pseudo-Voigt"
-                res = np.array([val, x])
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=cen,
-                    amp=amp,
-                    fwg=fwg,
-                    fwl=fwl,
-                    fwh=None,
-                    gam=None,
-                    dec=None,
-                    exp=None,
-                    slp=None,
-                    con=None,
-                    sig=None,
-                )
-                index += 1
-                # printf(tmp,x, val,'PS-Voigt',index,cen,amp,fwg,fwl,0,0,0,0,0,0,0)
-
-        if "expo" in mode:
-            if "dec" in mode:
-                dec = result.params[mode]
-            if "amp" in mode:
-                amp = result.params[mode]
-            if "con" in mode:
-                con = result.params[mode]
-
-                val = np.nan_to_num(exponential(x, amp, dec, con))
-                plt.plot(x, val, color="C3", ls="-.")
-                out.append(val)
-                # Exporting
-                model = "Exponential"
-                res = np.array([val, x])
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=None,
-                    amp=amp,
-                    fwg=None,
-                    fwl=None,
-                    fwh=None,
-                    gam=None,
-                    dec=dec,
-                    exp=None,
-                    slp=None,
-                    con=con,
-                    sig=None,
-                )
-                index += 1
-                # printf(tmp,x, val,'Exponential',index,0,amp,0,0,0,0,dec=dec,0,0,con,0)
-
-        if "powr" in mode:
-            if "ord" in mode:
-                exp = result.params[mode]
-            if "amp" in mode:
-                amp = result.params[mode]
-            if "con" in mode:
-                con = result.params[mode]
-
-                val = np.nan_to_num(powerlaw(x, amp, exp, con))
-                plt.plot(x, val, color="C3", ls="-.")
-                out.append(val)
-                # Exporting
-                model = "Power"
-                res = np.array([val, x])
-
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=None,
-                    amp=amp,
-                    fwg=None,
-                    fwl=None,
-                    fwh=None,
-                    gam=None,
-                    dec=None,
-                    exp=exp,
-                    slp=None,
-                    con=con,
-                    sig=None,
-                )
-                # printf(tmp,x, val,'Power',index,0,amp,0,0,0,0,0,exp=exp,0,con,0)
-
-        if "linr" in mode:
-            if "slp" in mode:
-                slp = result.params[mode]
-            if "con" in mode:
-                con = result.params[mode]
-                val = np.nan_to_num(linear(x, slp, con))
-                plt.plot(x, val, color="C3", ls="-.")
-                out.append(val)
-                # Exporting
-                model = "Linear"
-                res = np.array([val, x])
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=None,
-                    amp=None,
-                    fwg=None,
-                    fwl=None,
-                    fwh=None,
-                    gam=None,
-                    dec=None,
-                    exp=None,
-                    slp=slp,
-                    con=con,
-                    sig=None,
-                )
-                index += 1
-                # printf(tmp,x, val,'Linear',index,0,0,0,0,0,0,0,0,slp,con,0)
-
-        if "cons" in mode and "con" in mode:
-            con = result.params[mode]
-            val = np.nan_to_num(const(x, con))
-            plt.plot(x, val, color="C3", ls="-.")
-            out.append(val)
-            # Exporting
-            model = "Constant"
-            res = np.array([val, x])
-            df = panda_e(
-                df,
-                model,
-                index,
-                res,
-                cen=None,
-                amp=None,
-                fwg=None,
-                fwl=None,
-                fwh=None,
-                gam=None,
-                dec=None,
-                exp=None,
-                slp=None,
-                con=con,
-                sig=None,
-            )
-            index += 1
-            # printf(tmp,x, val,'Constant',index,0,0,0,0,0,0,0,0,0,con,0)
-
-        # elif 'stpl' in mode:
-        #    if 'cen' in mode: cen  = params[mode]
-        #    elif 'sig' in mode:
-        #        sig   = params[mode]
-        #        val += np.nan_to_num(step(x, cen, sig, form='linear'))
-        if "erf" in mode:
-            if "cen" in mode:
-                cen = result.params[mode]
-            if "sig" in mode:
-                sig = result.params[mode]
-            if "amp" in mode:
-                amp = result.params[mode]
-
-                val = np.nan_to_num(step(x, amp, cen, sig, form="erf"))
-                plt.plot(x, val, color="C6", ls="--")
-                out.append(val)
-                # Exporting
-                model = "Error-Step"
-                res = np.array([val, x])
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=cen,
-                    amp=amp,
-                    fwg=None,
-                    fwl=None,
-                    fwh=None,
-                    gam=None,
-                    dec=None,
-                    exp=None,
-                    slp=None,
-                    con=None,
-                    sig=sig,
-                )
-                index += 1
-                # printf(tmp,x, val,'Error-Fnc',index,cen,amp,0,0,0,0,0,0,0,0,sig=sig)
-
-        if "atan" in mode:
-            if "cen" in mode:
-                cen = result.params[mode]
-            if "sig" in mode:
-                sig = result.params[mode]
-            if "amp" in mode:
-                amp = result.params[mode]
-                val = np.nan_to_num(step(x, amp, cen, sig, form="atan"))
-                plt.plot(x, val, color="C6", ls="--")
-                out.append(val)
-                # Exporting
-                model = "ATangH-Step"
-                res = np.array([val, x])
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=cen,
-                    amp=amp,
-                    fwg=None,
-                    fwl=None,
-                    fwh=None,
-                    gam=None,
-                    dec=None,
-                    exp=None,
-                    slp=None,
-                    con=None,
-                    sig=sig,
-                )
-                index += 1
-                # printf(tmp,x, val,'ATangH-Fnc',index,cen,amp,0,0,0,0,0,0,0,0,sig=sig)
-
-        if "log" in mode:
-            if "cen" in mode:
-                cen = result.params[mode]
-            if "sig" in mode:
-                sig = result.params[mode]
-            if "amp" in mode:
-                amp = result.params[mode]
-                val = np.nan_to_num(step(x, amp, cen, sig, form="atan"))
-                plt.plot(x, val, color="C6", ls="--")
-                out.append(val)
-                # Exporting
-                model = "Log-Step"
-                res = np.array([val, x])
-
-                df = panda_e(
-                    df,
-                    model,
-                    index,
-                    res,
-                    cen=cen,
-                    amp=amp,
-                    fwg=None,
-                    fwl=None,
-                    fwh=None,
-                    gam=None,
-                    dec=None,
-                    exp=None,
-                    slp=None,
-                    con=None,
-                    sig=sig,
-                )
-                index += 1
-                # printf(tmp,x, val,'Log-Fnc',index,cen,amp,0,0,0,0,0,0,0,0,sig=sig)
-
-    pd.set_option(
-        "expand_frame_repr", args.disp, "display.float_format", lambda x: "%.3f" % x
-    )
-    # dates = pd.date_range('1/1/2000', periods=7)
-
-    print(df)
-    excle(filename + "_Export.xlsx", df)
-    # majorLocator = MultipleLocator(20)
-    # majorFormatter = FormatStrFormatter('%d')
-    # minorLocator = MultipleLocator(5)
-
-    # fig, ax = plt.subplots()
-
-    # ax.xaxis.set_major_locator(majorLocator)
-    # ax.xaxis.set_major_formatter(majorFormatter)
-
-    # for the minor ticks, use no labels; default NullFormatter
-
-    ax.xaxis.set_minor_locator(AutoMinorLocator())
-
-    plt.plot(x, y, color="C7", marker="+")
-    plt.plot(x, final, color="C1", ls="-.")
-    plt.tick_params(which="both", width=2)
-    plt.tick_params(which="major", length=7)
-    plt.tick_params(which="minor", length=4, color="0.75")
-    fig.savefig(filename + "_Fit.pdf", format="pdf")
-    cursor = Cursor(ax, useblit=True, color="C0", linewidth=1)
-    plt.show()
-    # tmp.close()
-    np.savetxt(filename + "_Fit.txt", np.transpose(out), delimiter="\t")
-    # high_res(x,final,result,filename)
-    # except:
-    #    pass
-
-
-def panda_h():
-    f_array = np.array([], dtype=np.float)
-    s_array = np.array([], dtype=np.str)
-    return pd.DataFrame(
-        {
-            "Model": s_array,
-            "Energy": f_array,
-            "stderr Energy": f_array,
-            "Amp": f_array,
-            "stderr Amp": f_array,
-            "Area": f_array,  #'stderr Area': f_array,
-            "FWHM (GAUSSIAN)": f_array,
-            "stderr FWHM (GAUSSIAN)": f_array,
-            "FWHM (LORENTZ)": f_array,
-            "stderr FWHM (LORENTZ)": f_array,
-            "FWHM (VOIGT)": f_array,
-            "stderr FWHM (VOIGT)": f_array,
-            "Fraction": f_array,
-            "stderr Fraction": f_array,
-            "Decay": f_array,
-            "stderr Decay": f_array,
-            "Power": f_array,
-            "stderr Power": f_array,
-            "Slope": f_array,
-            "stderr Slope": f_array,
-            "Constant": f_array,
-            "stderr Constant": f_array,
-            "Signum": f_array,
-            "stderr Signum": f_array,
-        }
-    )
-
-
-def panda_e(
-    file,
-    model,
-    index,
-    data,
-    cen=None,
-    amp=None,
-    fwg=None,
-    fwl=None,
-    fwh=None,
-    gam=None,
-    dec=None,
-    exp=None,
-    slp=None,
-    con=None,
-    sig=None,
-):
-    area = np.trapz(data[0], data[1])
-
-    if cen is None:
-        cen_fit, cen_error = np.array([0.0]), np.array([0.0])
-    else:
-        cen_fit, cen_error = np.round(cen._val, 8), np.array([0.0])
-    if amp is None:
-        amp_fit, amp_error = np.array([0.0]), np.array([0.0])
-    else:
-        amp_fit, amp_error = np.round(amp._val, 8), np.array([0.0])
-    if fwg is None:
-        fwg_fit, fwg_error = np.array([0.0]), np.array([0.0])
-    else:
-        fwg_fit, fwg_error = np.round(fwg._val, 8), np.array([0.0])
-    if fwl is None:
-        fwl_fit, fwl_error = np.array([0.0]), np.array([0.0])
-    else:
-        fwl_fit, fwl_error = np.round(fwl._val, 8), np.array([0.0])
-    if fwh is None:
-        fwh_fit, fwh_error = np.array([0.0]), np.array([0.0])
-    else:
-        fwh_fit, fwh_error = np.round(fwh._val, 8), np.array([0.0])
-    if gam is None:
-        gam_fit, gam_error = np.array([0.0]), np.array([0.0])
-    else:
-        gam_fit, gam_error = np.round(gam._val, 8), np.array([0.0])
-    if dec is None:
-        dec_fit, dec_error = np.array([0.0]), np.array([0.0])
-    else:
-        dec_fit, dec_error = np.round(dec._val, 8), np.array([0.0])
-    if exp is None:
-        exp_fit, exp_error = np.array([0.0]), np.array([0.0])
-    else:
-        exp_fit, exp_error = np.round(exp._val, 8), np.array([0.0])
-    if slp is None:
-        slp_fit, slp_error = np.array([0.0]), np.array([0.0])
-    else:
-        slp_fit, slp_error = np.round(slp._val, 8), np.array([0.0])
-    if con is None:
-        con_fit, con_error = np.array([0.0]), np.array([0.0])
-    else:
-        con_fit, con_error = np.round(con._val, 8), np.array([0.0])
-    if sig is None:
-        sig_fit, sig_error = np.array([0.0]), np.array([0.0])
-    else:
-        sig_fit, sig_error = np.round(sig._val, 8), np.array([0.0])
-    # if not (cen  is None): cen_fit, cen_error  = np.round(cen._val,8) , np.round(cen.stderr,8)
-    # if not (amp  is None): amp_fit, amp_error  = np.round(amp._val,8) , np.round(amp.stderr,8)
-    # if not (fwg  is None): fwg_fit, fwg_error  = np.round(fwg._val,8) , np.round(fwg.stderr,8)
-    # if not (fwl  is None): fwl_fit, fwl_error  = np.round(fwl._val,8) , np.round(fwl.stderr,8)
-    # if not (fwh  is None): fwh_fit, fwh_error  = np.round(fwh._val,8) , np.round(fwh.stderr,8)
-    # if not (gam  is None): gam_fit, gam_error  = np.round(gam._val,8) , np.round(gam.stderr,8)
-    # if not (dec  is None): dec_fit, dec_error  = np.round(dec._val,8) , np.round(dec.stderr,8)
-    # if not (exp  is None): exp_fit, exp_error  = np.round(exp._val,8) , np.round(exp.stderr,8)
-    # if not (slp  is None): slp_fit, slp_error  = np.round(slp._val,8) , np.round(slp.stderr,8)
-    # if not (con  is None): con_fit, con_error  = np.round(con._val,8) , np.round(con.stderr,8)
-    # if not (sig  is None): sig_fit, sig_error  = np.round(sig._val,8) , np.round(sig.stderr,8)
-
-    tmp = pd.DataFrame(
-        {
-            "Model": model,
-            "Energy": cen_fit,
-            "stderr Energy": cen_error,
-            "Amp": amp_fit,
-            "stderr Amp": amp_error,
-            "Area": np.round(area, 8),  #'stderr Area': f_array,
-            "FWHM (GAUSSIAN)": fwg_fit,
-            "stderr FWHM (GAUSSIAN)": fwg_error,
-            "FWHM (LORENTZ)": fwl_fit,
-            "stderr FWHM (LORENTZ)": fwl_error,
-            "FWHM (VOIGT)": fwh_fit,
-            "stderr FWHM (VOIGT)": fwh_error,
-            "Fraction": gam_fit,
-            "stderr Fraction": gam_error,
-            "Decay": dec_fit,
-            "stderr Decay": dec_error,
-            "Power": exp_fit,
-            "stderr Power": exp_error,
-            "Slope": slp_fit,
-            "stderr Slope": slp_error,
-            "Constant": con_fit,
-            "stderr Constant": con_error,
-            "Signum": sig_fit,
-            "stderr Signum": sig_error,
-        },
-        index=[index],
-    )
-
-    # merged = file.append(tmp)
-    # cols = list(file) + list(tmp)
-    # merged.columns = cols
-
-    index = [
-        "Model",
-        "Energy",
-        "stderr Energy",
-        "Amp",
-        "stderr Amp",
-        "Area",  #'stderr Area': f_
-        "FWHM (GAUSSIAN)",
-        "stderr FWHM (GAUSSIAN)",
-        "FWHM (LORENTZ)",
-        "stderr FWHM (LORENTZ)",
-        "FWHM (VOIGT)",
-        "stderr FWHM (VOIGT)",
-        "Fraction",
-        "stderr Fraction",
-        "Decay",
-        "stderr Decay",
-        "Power",
-        "stderr Power",
-        "Slope",
-        "stderr Slope",
-        "Constant",
-        "stderr Constant",
-        "Signum",
-        "stderr Signum",
-    ]
-
-    file = file.append(tmp)
-
-    return file[index]
-
-
-def excle(filename, file):
-    # Create a Pandas Excel writer using XlsxWriter as the engine.
-    # writer = pd.ExcelWriter(filename, engine="xlsxwriter")
-
-    # Convert the dataframe to an XlsxWriter Excel object.
-    file.to_excel(filename, sheet_name="Fit-Results")
-
-    # Close the Pandas Excel writer and output the Excel file.
-    # writer.save()
-
-
-def pv_cof(fwhm_g: float, fwhm_l: float) -> Tuple[float, float]:
-    """
-    Calculating the effectiv fwhm of the pseudo voigt profile and the fraction coefficient for n
-    """
-    f = np.power(
-        fwhm_g ** 5
-        + 2.69269 * fwhm_g ** 4 * fwhm_l
-        + 2.42843 * fwhm_g ** 3 * fwhm_l ** 2
-        + 4.47163 * fwhm_g ** 2 * fwhm_l ** 3
-        + 0.07842 * fwhm_g * fwhm_l ** 4
-        + fwhm_l ** 5,
-        0.25,
-    )
-    n = (
-        1.36603 * (fwhm_l / f)
-        - 0.47719 * (fwhm_l / f) ** 2
-        + 0.11116 * (fwhm_l / f) ** 3
-    )
-    return (f, n)
-
-
-def gaussian(
-    x: np.array, amplitude: float = 1.0, center: float = 0.0, fwhm: float = 1.0
-) -> np.array:
-    """Return a 1-dimensional Gaussian function.
-
-    gaussian(x, amplitude, center, sigma) =
-        (amplitude/(sq2pi*sigma)) * exp(-(1.0*x-center)**2 / (2*sigma**2))
-
-    """
-    sigma = fwhm / Constants.sig2fwhm
-    return (amplitude / (Constants.sq2pi * sigma)) * np.exp(
-        -((1.0 * x - center) ** 2) / (2 * sigma ** 2)
-    )
-
-
-def lorentzian(
-    x, amplitude: float = 1.0, center: float = 0.0, fwhm: float = 1.0
-) -> np.array:
-    """Return a 1-dimensional Lorentzian function.
-
-    lorentzian(x, amplitude, center, sigma) =
-        (amplitude/(1 + ((1.0*x-center)/sigma)**2)) / (pi*sigma)
-
-    """
-    sigma = fwhm / 2.0
-    return (amplitude / (1 + ((1.0 * x - center) / sigma) ** 2)) / (np.pi * sigma)
-
-
-def voigt(
-    x: np.array, center: float = 0.0, fwhm: float = 1.0, gamma: float = None
-) -> np.array:
-    """Return a 1-dimensional Voigt function.
-
-    voigt(x, amplitude, center, sigma, gamma) =
-        amplitude*wofz(z).real / (sigma*sq2pi)
-
-    see http://en.wikipedia.org/wiki/Voigt_profile
-
-    """
-    sigma = fwhm / 3.60131
-    if gamma is None:
-        gamma = sigma
-    z = (x - center + 1j * gamma) / (sigma * Constants.s2)
-    return wofz(z).real / (sigma * Constants.sq2pi)
-
-
-# def pseudovoigt(x, amplitude=1.0, center=0.0, sigma=1.0, fraction=0.5):
-#    """Return a 1-dimensional pseudo-Voigt function.
-#
-#    pseudovoigt(x, amplitude, center, sigma, fraction) =
-#       amplitude*(1-fraction)*gaussion(x, center, sigma_g) +
-#       amplitude*fraction*lorentzian(x, center, sigma)
-#
-#    where sigma_g (the sigma for the Gaussian component) is
-#
-#        sigma_g = sigma / sqrt(2*log(2)) ~= sigma / 1.17741
-#
-#    so that the Gaussian and Lorentzian components have the
-#    same FWHM of 2*sigma.
-#
-#    """
-#    sigma_g = sigma / np.sqrt(2*log2)
-#    return ((1-fraction)*gaussian(x, amplitude, center, sigma_g) +
-#            fraction*lorentzian(x, amplitude, center, sigma))
-
-
-def pseudovoigt(
-    x: np.array,
-    amplitude: float = 1.0,
-    center: float = 0.0,
-    fwhm_g: float = 1.0,
-    fwhm_l: float = 1.0,
-) -> np.array:
-
-    # PSVOIGT This is a psuedo voigt function
-    #   This is programmed according to wikipidia and they cite:
-    #   Ida, T and Ando, M and Toraya, H (2000).
-    # "Extended pseudo-Voigt function for approximating the Voigt profile".
-    # Journal of Applied Crystallography 33 (6): 1311-1316.
-
-    # sigma_g = sigma / np.sqrt(2*log2)
-    f, n = pv_cof(fwhm_g, fwhm_l)
-
-    # f = np.power(fwhm_g**5 +2.69269*fwhm_g**4*fwhm_l+2.42843*fwhm_g**3*fwhm_l**2+
-    #             4.47163*fwhm_g**2*fwhm_l**3+0.07842*fwhm_g*fwhm_l**4+fwhm_l**5,0.25)
-    # n = 1.36603*(fwhm_l/f) - 0.47719*(fwhm_l/f)**2 +0.11116*(fwhm_l/f)**3
-    return n * lorentzian(x, amplitude, center, fwhm_l) + (1 - n) * gaussian(
-        x, amplitude, center, fwhm_g
-    )
-
-
-def exponential(
-    x: np.array, amplitude: float = 1.0, decay: float = 1.0, intercept: float = 0.0
-) -> np.array:
-    """Return an exponential function.
-
-    x -> amplitude * exp(-x/decay)
-
-    """
-    return amplitude * np.exp(-x / decay) + intercept
-
-
-def powerlaw(
-    x: np.array, amplitude: float = 1.0, exponent: float = 1.0, intercept: float = 0.0
-) -> np.array:
-    """Return the powerlaw function.
-
-    x -> amplitude * x**exponent
-
-    """
-
-    return amplitude * np.power(x, exponent) + intercept
-
-
-def linear(x: np.array, slope: float = 1.0, intercept: float = 0.0) -> np.array:
-    """Return a linear function.
-
-    x -> slope * x + intercept
-
-    """
-    return slope * x + intercept
-
-
-def constant(x: np.array, amplitude: float = 1.0) -> np.array:
-    """Return a cosntant function.
-
-    x -> constant
-
-    """
-    return np.linspace(amplitude, amplitude, len(x))
-
-
-def step(
-    x: np.array,
-    amplitude: float = 1.0,
-    center: float = 0.0,
-    sigma: float = 1.0,
-    form: str = "linear",
-) -> np.array:
-    """Return a step function.
-
-    starts at 0.0, ends at amplitude, with half-max at center, and
-    rising with form:
-      'linear' (default) = amplitude * min(1, max(0, arg))
-      'atan', 'arctan'   = amplitude * (0.5 + atan(arg)/pi)
-      'erf'              = amplitude * (1 + erf(arg))/2.0
-      'logistic'         = amplitude * [1 - 1/(1 + exp(arg))]
-
-    where arg = (x - center)/sigma
-
-    """
-    if abs(sigma) < 1.0e-13:
-        sigma = 1.0e-13
-
-    out = (x - center) / sigma
-    if form == "erf":
-        out = 0.5 * (1 + erf(out))
-    elif form.startswith("logi"):
-        out = 1.0 - 1.0 / (1.0 + np.exp(out))
-    elif form in {"atan", "arctan"}:
-        out = 0.5 + np.arctan(out) / np.pi
-    else:
-        out[np.where(out < 0)] = 0.0
-        out[np.where(out > 1)] = 1.0
-    return amplitude * out
 
 
 if __name__ == "__main__":
